% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ridge.R
\name{biasRidge}
\alias{biasRidge}
\title{Bias of Ridge Estimator}
\usage{
biasRidge(lambda, beta, XtX, penalize = c(0, rep(1, ncol(XtX) - 1)),
  ind = 1, XtXlamIinv = NULL)
}
\arguments{
\item{lambda}{Penalty parameter (should have length 1).}

\item{beta}{True parameter values. Either a vector of length \code{p} or a 
\code{p} x \code{d} matrix.}

\item{XtX}{Cross product of design matrix}

\item{penalize}{Vector of penalty factors. See \code{\link{estRidge}} for more information.}

\item{ind}{Numerical or logical vector indicating which elements
should be returned. Defaults to the first element.}

\item{XtXlamIinv}{Optional explicit value of \code{(XtX + diag(lambda*penalize))^(-1)}. Pre-computing this may save time for simulations.}
}
\description{
Computes the analytic bias of ridge estimator
}
\details{
If \code{beta} is provided as a matrix, this will treat
each column of \code{beta} as a different true parameter vector 
and return a matrix of bias values (or a vector, if \code{ind} has length 1).
}
\author{
Joshua Keller
}
\seealso{
\code{\link{mseRidge}, \link{varRidge}}
}

