% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eshrink.R
\docType{package}
\name{eshrink-package}
\alias{eshrink-package}
\alias{eshrink}
\title{Shrinkage Estimators for Regression}
\description{
Computes shrinkage estimators for regression problems. Selects
penalty parameter by minimizing bias and variance in the effect estimate,
where bias and variance are estimated from the posterior predictive
 distribution. See Keller and Rice (2017) <doi:10.1093/aje/kwx225> for more 
details.
}
