% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ridge.R
\name{estRidge}
\alias{estRidge}
\title{Estimate Coefficients for Ridge Regression}
\usage{
estRidge(lambda, X, y, penalize, XtX = crossprod(X))
}
\arguments{
\item{lambda}{Ridge penalty factor}

\item{X}{Design matrix for the regression. Assumed to contain only numeric values, so 
any factors should be coded according to desired contrast (e.g., via \code{\link{model.matrix}})}

\item{y}{Outcome vector. Unless \code{X} contains an intercept column, this should typically be centered.}

\item{penalize}{Vector giving penalty structure. Values must be in [0, 1].}

\item{XtX}{the cross product of the design matrix. Computed from
\code{X} if not provided.}
}
\description{
Computes a vector of regression coefficients for a provided ridge penalty.
}
\details{
The input \code{penalize} is a vector of penalty factors, 
such that the penalty for covariate j is \code{lambda*penalize[j]}.
Although its primary purpose is for indicating which variables to penalize (1)
and which to not penalize (0), fractional values between 0 and 1 are accepted.
}
\author{
Joshua Keller
}
\seealso{
\code{\link{festRidge}}, \code{\link{mseRidge}}
}

