% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_aux.R
\name{getMSE}
\alias{getMSE}
\title{Mean Squared Error of a Vector}
\usage{
getMSE(x, truth)
}
\arguments{
\item{x}{vector of estimates}

\item{truth}{vector of true parameters}
}
\description{
Compute the Mean Squared Error of a vector
}
\details{
Details here!
}

