% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_lasso.R
\name{lassofEst}
\alias{lassofEst}
\title{Compute Lasso Estimator for simulated Data}
\usage{
lassofEst(X, y, type = c("fMSE", "fMBV", "all"), lseq, B = 500,
  penalize = NULL, rescale.lambda = TRUE, scale = FALSE,
  returnMSE = FALSE, se.version = c("varExp", "full", "both"), postsamp,
  returnPS = FALSE, nPost = 1000, ...)
}
\arguments{
\item{X}{Design matrix of regression problem}

\item{y}{outcome vector. Typically centered.}

\item{type}{Which type of estimate should be computed? See details.}

\item{lseq}{Sequence of penalty values to consider.}

\item{B}{Number of future datasets to simulate for each point in posterior sample.}

\item{penalize}{Vector giving penalty structure. Supplied to glmnet as `\code{penalty.factor}'.}

\item{rescale.lambda}{Should lambda be rescaled to account for the 
default re-scaling done by glmnet?}

\item{scale}{Logical indicating whether the design matrix X be scaled. See details.}

\item{returnMSE}{Logical indicating whether mse object should be returned.}

\item{se.version}{Which version of Standard errors to report?}

\item{postsamp}{List containing posterior sample (from \code{samplePosterior}). If
missing, then a posterior sample is drawn.  Currently checks on the provided
\code{postsamp} are limited, so use with caution.  Designed to facilitate
simualtions or other scenarios where it may be pre-computed.}

\item{returnPS}{logical indicating whether or not the full posterior sample should
be included in output.}

\item{nPost}{Size of posterior sample to compute}

\item{...}{Other arguments passed to \code{samplePosterior}}

\item{lambda}{Penalty factor to be applied}
}
\description{
To be added...
}
\details{
To be added...
}
\seealso{
\code{\link{ridgefEst}}, \code{\link{simLassoEst}}
}

