% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ridge.R
\name{mseRidge}
\alias{mseRidge}
\title{Compute MSE for Ridge Estimator}
\usage{
mseRidge(lambda, XtX, beta, sigma2, penalize, ind = 1, XtXlamIinv = NULL)
}
\arguments{
\item{lambda}{Vector of shrinkage penalty parameters.}

\item{XtX}{The cross product of the design matrix.}

\item{beta}{True value of the regression parameters}

\item{sigma2}{True value of the variance parameter}

\item{penalize}{See \code{\link{estRidge}}}

\item{ind}{Numeric or logical vector indicating which
element for which MSE should be returned}

\item{XtXlamIinv}{See details.}
}
\description{
Computes the mean-squared error (MSE) for ridge estimators given different
	values of the true \code{beta} and \code{sigma2} parameters.
}
\details{
The bias and variance are computed using the functions
	\code{\link{biasRidge}} and \code{\link{varRidge}}. 

	Explicit value(s) of (X'X + diag(penalty))^{-1} can be provided
	through the argument \code{XtXlamIinv}. If length of \code{lambda} is greater than 1,
	this should be provided as a list.
}
\author{
Joshua Keller
}
\seealso{
\code{\link{varRidge}, \link{biasRidge}}
}

