% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ridge.R
\name{ridgeBias}
\alias{ridgeBias}
\title{Get Bias of Ridge Estimator}
\usage{
ridgeBias(lambda, beta, XtX, penalize = c(0, rep(1, ncol(XtX) - 1)),
  ind = 1, XtXlamIinv = NULL)
}
\arguments{
\item{lambda}{Penalty parameter (should have length 1).}

\item{beta}{True parameter vector. Can also be p x n matrix.}

\item{XtX}{cross product of design matrix}

\item{penalize}{vector giving what to penalize. See \code{\link{ridgeEst}} for more information.}

\item{ind}{Numerical or logical vector indicating which elements
should be returned. Defaults to the first element.}

\item{XtXlamIinv}{explicit value of (X'X + diag(penalty))^{-1}.  Useful
for simulations to save computation.}
}
\description{
Computes the (analytic) bias of Ridge estimator
}
\details{
If \code{beta} is provided as a matrix, this will
}
\author{
Joshua Keller
}
\seealso{
\code{\link{ridgeMSE}, \link{ridgeVar}}
}

