% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ridge.R
\name{ridgeEst}
\alias{ridgeEst}
\title{Compute Ridge Estimate}
\usage{
ridgeEst(lambda, X, y, penalize, XtX = crossprod(X))
}
\arguments{
\item{lambda}{Penalty factor to be applied}

\item{X}{Design matrix of regression problem}

\item{y}{outcome vector. Typically centered.}

\item{penalize}{Vector giving penalty structure. Values must be in [0, 1].}

\item{XtX}{the cross product of the design matrix. Will be computed from
\code{X} if not provided.}
}
\description{
Computes a ridge estimate for a given regression problem
}
\details{
Computes the ridge estimator from the regression of y on X, with penalty (vector!) penalize.

The input \code{penalize} provides a vector of penalty factors, 
such that the penalty for covariate j is lambda*penalize[j].
Primary purpose is for indicating which variable to penalize (1)
and which to not (0), although more flexible penalties that include
}
\seealso{
\code{\link{ridgefEst}}, \code{\link{ridgeMSE}}
}

