% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ridge.R
\name{ridgefEst}
\alias{ridgefEst}
\title{Compute `Future MSE' Ridge Estimate}
\usage{
ridgefEst(X, y, type = c("fMSE", "fMBV", "all"), lseq, scale = FALSE,
  returnMSE = FALSE, postsamp, returnPS = FALSE, nPost = 1000,
  se = TRUE, XtXlamIinv = NULL, ...)
}
\arguments{
\item{X}{Design matrix of regression problem}

\item{y}{outcome vector. Typically centered.}

\item{type}{The type of estimator. See details.}

\item{lseq}{Sequence of penalty values to consider.}

\item{scale}{Logical indicating whether the design matrix X be scaled. See details.}

\item{returnMSE}{Logical indicating whether mse object should be returned.}

\item{postsamp}{List containing posterior sample (from \code{samplePosterior}). If
missing, then a posterior sample is drawn.  Currently checks on the provided
\code{postsamp} are limited, so use with caution.  Designed to facilitate
simualtions or other scenarios where it may be pre-computed.}

\item{returnPS}{logical indicating whether or not the full posterior sample should
be included in output.}

\item{nPost}{Size of posterior sample to compute}

\item{se}{logical indicating whether or not standard errors should be returned.}

\item{XtXlamIinv}{explicit value of (X'X + diag(penalty))^{-1}.  Useful
for simulations to save computation.}

\item{...}{Other arguments passed to \code{samplePosterior}}
}
\description{
Computes a ridge estimate for a given regression problem, with penalty parameter chosen to minimize future MSE.
}
\details{
To be added.....

	To balance the influence of covariates, it is recommended
	that the design matrix be standardized.  This can be done by
	the user or via the argument \code{scale}.  If \code{scale=TRUE},
	then coefficent and standard error estimates are back-transformed.
}
\seealso{
\code{\link{lassofEst}}, \code{\link{ridgeMSE}}
}

