% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ridge.R
\name{ridgeMSE}
\alias{ridgeMSE}
\title{Compute MSE for Ridge Estimator}
\usage{
ridgeMSE(lambda, XtX, beta, sigma2, ind = 1, XtXlamIinv = NULL)
}
\arguments{
\item{lambda}{Shrinkage penalty parameters (vector).}

\item{XtX}{the cross product of the design matrix.}

\item{beta}{true value of the regression parameters}

\item{sigma2}{true value of the variance parameter}

\item{ind}{numeric or logical vector indicating which
element for which MSE should be returned}

\item{XtXlamIinv}{explicit value of (X'X + diag(penalty))^{-1}.  Useful
for simulations to save computation.}
}
\description{
Computes the MSE for ridge estimators given different
	values of the true \code{beta} and \code{sigma2} coefficients.
}
\details{
The bias and variance are computed using the functions
	\code{\link{ridgeBias}} and \code{\link{ridgeVar}}. 

	The \code{XtXlamIinvB} and \code{returmMSE} arguments were designed
	for simulations, where pre-computation of matrices and reducing
	output provides computational benefits.  These are typically
	not needed for analysis of a single dataset.
}

