% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ridge.R
\name{ridgeVar}
\alias{ridgeVar}
\title{Get Ridge variance}
\usage{
ridgeVar(lambda, XtX, penalize, sigma2 = 1, ind = 1, XtXlamIinv = NULL)
}
\arguments{
\item{lambda}{penalty parameter}

\item{XtX}{cross product of design matrix}

\item{penalize}{vector giving what to penalize}

\item{sigma2}{The true variance parameter}

\item{ind}{Numerical or logical vector indicating which elements
of the variance matrix should be returned. Defaults to the
(1,1) element}

\item{XtXlamIinv}{explicit value of (X'X + diag(penalty))^{-1}.  Useful
for simulations to save computation.}
}
\description{
To be added..
}
\details{
See \code{\link{ridgeEst}} for description of \code{penalize} parameter.
}
\author{
Joshua Keller
}
\seealso{
\code{\link{ridgeMSE}, \link{ridgeBias}}
}

