% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bayes.R
\name{samplePosterior}
\alias{samplePosterior}
\title{Posterior Sample for Bayesian Linear Regression}
\usage{
samplePosterior(X, y, n, a0 = 1, b0 = 5e-05, v0inv = 1/1000, mu0 = 0,
  hyperPost = FALSE, intercept = FALSE)
}
\arguments{
\item{X}{A data matrix of size \code{n} by \code{p}.}

\item{y}{Outcome variable}

\item{n}{Size of posterior sample}

\item{a0, b0}{Hyperparameters for inverse gamma
distribution of the error varaince.}

\item{v0inv}{Prior precision for the error term.  Either a single value
to be repeated in a diagonal precision matrix, or a \code{p} by \code{p}
matrix.}

\item{mu0}{Prior mean.  Defaults to zero vector.}

\item{hyperPost}{Logical indicating whether the posterior hyperparameters are returned.}

\item{intercept}{Logical indicating whether an intercept is included in the model.
If \code{FALSE}, then \code{y} is centered prior to computations.}
}
\value{
A list containing the following elements:
\item{sigma2}{A vector containing the posterior sample of \eqn{sigma^2} values.}
\item{beta}{Matrix containing the posterior sample of \eqn{beta} values.}
\item{postMu}{Vector containing the posterior mean (if \code{moments=TRUE}).}
\item{postV}{Matrix giving the posterior mean (if \code{moments=TRUE}).}
\item{wSS}{Mean within-cluster sum-of-squares}
}
\description{
Computes and provides a sample from the posterior
	distribution of parameters from a Bayesian Linear regression
	model.
}
\details{
Assumes a normal-inverse-gamma model, where the distribution
	of the regression coefficients is conditional on \eqn{sigma^2}.
}
\author{
Joshua Keller
}

