% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ridge.R
\name{varRidge}
\alias{varRidge}
\title{Variance of Ridge Estimator}
\usage{
varRidge(lambda, XtX, penalize, sigma2 = 1, ind = 1, XtXlamIinv = NULL)
}
\arguments{
\item{lambda}{Penalty parameter.}

\item{XtX}{Cross product of design matrix}

\item{penalize}{Vector giving what to penalize}

\item{sigma2}{The true variance parameter}

\item{ind}{Numerical or logical vector indicating which elements
of the variance matrix diagonal should be returned. Defaults to the
(1,1) element}

\item{XtXlamIinv}{Explicit value of \eqn{(X'X + diag(penalty))^(-1)},
where \code{penalty=lambda*penalize}. Useful
for simulations to save computation.}
}
\description{
Computes the variance matrix of a ridge estimator, conditional on the penalty and global variance \code{sigma2}.
}
\details{
See \code{\link{estRidge}} for description of \code{penalize} parameter.
}
\author{
Joshua Keller
}
\seealso{
\code{\link{mseRidge}, \link{biasRidge}}
}

